12-byte SAT support for libatasmart
Hi Lennart,

My USB enclosure (a Maxtor basics of some description) supports SAT, but
only the 12-byte version. I've included a patch that implements it.
RIght now it tries the 16 byte one and then falls back, which is
possibly not ideal - there's a significant pause before the 16 byte one
finishes and the 12 byte one is tried. I'd reverse the order, but I've
no idea if that would confuse other devices. Are we able to check for
which commands the firmware claims to support?

Signed-off-by: Lennart Poettering <lennart@poettering.net>