Get rid of qurik 9_UNKNOWN again
Apparently Seagate seems to use the MSB's of the raw data of the
power-on-time field for something else on some disks, so let's just look
at the lower 32bits of that field and we get a value that works on all
disks including those from Seagate. 32bit should be more than enough to
encode the power-on-time for almost all disks so this should not be much
of a limitation.