libatasmart patch
hi

I have made a patch for libatasmart to add support for attribute 184
and 188. just adds the name to them.
Anyway I thought you might want to take a look at it.

Iv attached the patch against the git master and here is the link to
the Ubuntu's bug/patch
https://bugs.launchpad.net/libatasmart/+bug/521864

Thanks
Guy Taylor